feature: added modal
used parallel routes and intercepting routes to add a modal that works on all pages
