+++ src/app/(app)/@modal/default.tsx
@@ -1,0 +1,3 @@
+export default function Default() {
+    return null
+  }
