+++ src/components/feedbackPage/Modal.tsx
@@ -1,0 +1,23 @@
+"use client";
+import { useRouter } from "next/navigation";
+
+export function Modal({ children }: { children: React.ReactNode }) {
+  const router = useRouter();
+
+  const close = () => router.back();
+
+  return (
+    <div className="fixed inset-0 z-50 flex items-center justify-center">
+      <div className="absolute inset-0 bg-black opacity-50" onClick={close} />
+      <div className="relative z-40 w-full max-w-md rounded-2xl bg-white p-6 shadow-lg">
+        <button
+          onClick={close}
+          className="absolute top-3 right-3 text-gray-500 hover:text-gray-700"
+        >
+          ✕
+        </button>
+        {children}
+      </div>
+    </div>
+  );
+}
